requirements.txt
allure-pytest==2.13.4
python-docx==1.1.0
pytest==8.3.2
pytest-dotenv==0.5.2
selenium==4.21.0
webdriver-manager==4.0.1
